requirements.txt
aiofiles==22.1.0
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.1.0
asttokens==2.4.1
attrs==23.1.0
choicelib==0.1.5
colorama==0.4.6
dnspython==1.16.0
executing==2.0.1
frozenlist==1.4.0
icecream==2.1.3
idna==3.6
multidict==6.0.4
pydantic==1.10.13
Pygments==2.17.2
pymongo==3.11.0
python-dotenv==1.0.0
six==1.16.0
sniffio==1.3.0
typing_extensions==4.8.0
vbml==1.1.post1
vkbottle==4.3.12
vkbottle-types==5.131.146.16
watchfiles==0.18.1
yarl==1.9.4
